OnlineEducation/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-grappelli==4.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
mysql-connector-python==9.1.0
pillow==11.1.0
sqlparse==0.5.3
